kwork-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.27.0
asyncio-throttle==1.0.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
cryptography==42.0.8
redis==5.0.1
sqlite-utils==3.35.2
asyncpg==0.29.0
email-validator==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx-retry==2025.4.23
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==2.2.0
python-dotenv==1.0.0
playwright==1.42.0